requirements.txt
# Core dependencies
streamlit>=1.28.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0  # For EfficientNet and other vision models

# Image Processing
opencv-python>=4.8.0
albumentations>=1.3.0

# Generative AI
ollama>=0.1.0  # For local LLM (Llama, Mistral)
# openai>=1.0.0  # Uncomment if using OpenAI API
# anthropic>=0.7.0  # Uncomment if using Anthropic API

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
huggingface_hub>=0.20.0  # For downloading HuggingFace datasets

# YOLO multi-food detection
ultralytics>=8.0.0  # For YOLOv8 multi-food detection

# Optional: For model optimization
# onnx>=1.14.0
# onnxruntime>=1.15.0

